# docs/openapi.yaml
openapi: 3.1.0
info:
  title: zkAPI v2
  version: 2.0.0-testnet
  description: Local compatibility API plus the private Groth16 protocol API.
paths:
  /v1/chat/completions:
    post:
      summary: OpenAI-compatible chat through the local wallet daemon
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true}
      responses:
        "200":
          description: OpenAI-compatible response
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
  /v1/responses:
    post:
      summary: OpenAI Responses-compatible request through the local wallet daemon
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true}
      responses:
        "200": {description: Responses-compatible result}
...